make throwing class easier to use
In hindsight, using templates to set after how many instantiations it
will throw was really annoying to use. So I've opted for a simple
function to make the next instantation/copy throw. I've also added
some data to the class so I can check if it did not get corrupted